Introduce segmentation orchestrator and modularize worker architecture with a new BaseWorker.
