+++ workers/environment.yml
@@ -3,11 +3,11 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.12
+  - python=3.10
   - ffmpeg
   - librosa
-  - numpy
-  - scipy
+  - numpy<1.24
+  - scipy<1.9.0
   - numba
   - scikit-learn
   - pika
