+++ workers/segmenters/custom_worker.py
@@ -1,6 +1,5 @@
 import os
-import time
-from workers.base import BaseWorker
+from workers.BaseWorker import BaseWorker
 from shared.logger import get_logger
 
 logger = get_logger()
